feat: improve code with valgrind

+++ src/main.cpp
@@ -20,10 +20,11 @@ int main(int argc, char *argv[]) {
       std::string exp_values = expression_values;
       std::string exp = expression;
       BinaryTree tree(exp_values, exp);
-      unsigned long long int lowerbound =
+      unsigned int lowerbound =
           find_next_quantificator_pos_after_index(exp_values, 0);
 
       std::cout << tree.evaluateTree(lowerbound) << std::endl;
+
       break;
     }
   } catch (const std::exception &e) {
